cordillera: nest client API as Http2.Endpoint / Http2.Client
`H2Endpoint` and the `Http2Client` given occupied the top-level namespace
with the generic word "client". Move them into the existing `Http2`
object as `Http2.Endpoint` and `Http2.Client` (alongside `Frame`,
`Setting`, `ErrorCode`, …), so the HTTP/2 client surface is grouped the
same way as the wire vocabulary and stays out of the way.
