--- lib/cordillera/src/core/cordillera.Http2Client.scala
@@ -1,66 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.54.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package cordillera
-
-import anticipation.*
-import coaxial.*
-import contingency.*
-import gossamer.*
-import parasite.*
-import prepositional.*
-import telekinesis.*
-
-// A cleartext-h2c endpoint: a connectable address plus the `:authority` to send.
-// Used as the `Target` of the HTTP/2 `HttpClient` given, distinct from the
-// `DomainSocket` target telekinesis binds to its HTTP/1.1 client.
-case class H2Endpoint[endpoint: Connectable as connectable](endpoint: endpoint, authority: Text):
-  def connect(): Duplex = connectable.connect(endpoint)
-
-object Http2Client:
-  // An `HttpClient` that speaks HTTP/2 (prior-knowledge h2c) to an `H2Endpoint`.
-  // It captures the ambient `Monitor`/`Codicil` from this given's context — the
-  // `H2Connection`'s daemons need them — so it can only be summoned inside a
-  // `supervise` scope. A fresh connection is opened per request for now; pooling
-  // is a later refinement.
-  given http2: [endpoint] => (Monitor, Codicil, Tactic[Http2Error], Tactic[AsyncError])
-  =>  HttpClient onto H2Endpoint[endpoint] =
-
-    new HttpClient:
-      type Target = H2Endpoint[endpoint]
-
-      def request(request: Http.Request, target: H2Endpoint[endpoint])
-      :   Http.Response logs HttpEvent =
-
-        val connection = H2Connection(target.connect())
-        connection.start()
-        connection.fetch(request, t"http", target.authority)(1)
